+++ android/impl.nix
@@ -21,7 +21,7 @@ let overrideAndroidCabal = package: overrideCabal package (drv: {
       buildInputs = [ androidenv.androidPkgs_9_0.androidsdk ];
     } "";
     buildGradleApp = import ./build-gradle-app.nix {
-      inherit (nixpkgs) stdenv jdk gnumake gawk file runCommand
+      inherit (nixpkgs) stdenv lib jdk gnumake gawk file runCommand
                      which gradle fetchurl buildEnv;
       inherit androidenv;
     };
@@ -48,10 +48,10 @@ in {
               myNixpkgs = nixpkgsCross.android.aarch64;
               myHaskellPackages = ghcAndroidAarch64;
             };
-            # "armeabi-v7a" = {
-            #   myNixpkgs = nixpkgsCross.android.aarch32;
-            #   myHaskellPackages = ghcAndroidAarch32;
-            # };
+            "armeabi-v7a" = {
+              myNixpkgs = nixpkgsCross.android.aarch32;
+              myHaskellPackages = ghcAndroidAarch32;
+            };
           };
           abiVersions = attrNames appSOs;
       in nixpkgs.runCommand "android-app" {